backend-py/requirements.txt
# --- Setup & run instructions ---
# 1. python3 -m venv .venv
# 2. source .venv/bin/activate
# 3. pip install -r requirements.txt
# 4. uvicorn main:app --reload --host 0.0.0.0 --port 8000  (run from backend-py/)
#    - or: uvicorn main:app --app-dir backend-py --reload --host 0.0.0.0 --port 8000 (from repo root)
# 5. Ensure backend-py/.env has valid REDIS_* settings or start a local Redis on localhost:6379

# FastAPI and server
fastapi
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Redis
redis==5.2.0

# LangChain and Anthropic
langchain==0.3.7
langchain-anthropic==0.3.3
langchain-core==0.3.15
anthropic==0.39.0

# HTTP client (for Sanity CMS)
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
